MIX-284 refactor: split blurchette game.tsx into feature sub-components
